Update rest_api_base.py
TableauRestApiBase36 does not initialise TableauRestAPIBase (commented out) I'm not 100% sure why, assume it has something to do with the implementation of personal access tokens. 
add_user_by_username() in user.py validates the specified site_role against the tuple of allowed site_roles as defined in TableauRestAPIBase which doesn't get initialised for versions after 3.6. Adding this tuple to TableauRestApiBase36 
Would be good to understand better why the chain of inheritance is broken in 3.6?